+++ src/pypdfium2/__init__.py
@@ -1,7 +1,6 @@
 # SPDX-FileCopyrightText: 2026 geisserml <geisserml@gmail.com>
 # SPDX-License-Identifier: Apache-2.0 OR BSD-3-Clause
 
-print("top-level __init__.py")  # XXX
 import pypdfium2._library_scope
 from pypdfium2.version import *
 from pypdfium2._helpers import *
