+++ src/pypdfium2_cli/_parsers.py
@@ -1,40 +1,10 @@
 # SPDX-FileCopyrightText: 2026 geisserml <geisserml@gmail.com>
 # SPDX-License-Identifier: Apache-2.0 OR BSD-3-Clause
 
-import os
 import sys
-import logging
-import warnings
 import argparse
 from pathlib import Path
 import pypdfium2._helpers as pdfium
-import pypdfium2.internal as pdfium_i
-
-
-def setup_logging():
-    
-    print("Setting up logging")  # XXX
-    
-    # could also pass through the log level by parameter, but using an env var seemed easiest for now
-    debug_autoclose = bool(int( os.environ.get("DEBUG_AUTOCLOSE", 0) ))
-    debug_sysfonts = bool(int( os.environ.get("DEBUG_SYSFONTS", 0) ))
-    debug_unsupported = bool(int( os.environ.get("DEBUG_UNSUPPORTED", 1) ))
-    loglevel = getattr(logging, os.environ.get("PYPDFIUM_LOGLEVEL", "debug").upper())
-    
-    pdfium_i.DEBUG_AUTOCLOSE.value = debug_autoclose
-    lib_logger = logging.getLogger("pypdfium2")
-    lib_logger.addHandler(logging.StreamHandler())
-    lib_logger.setLevel(loglevel)
-    
-    warnings.simplefilter("always")
-    logging.captureWarnings(True)
-    
-    if debug_unsupported:
-        pdfium.PdfUnspHandler().setup()
-    
-    print("Installing sysfontinfo...")  # XXX
-    if debug_sysfonts:
-        sysfont_listener = pdfium.PdfSysfontListener()
 
 
 def parse_numtext(numtext):
